Allow to modify tables restricted with a row filter
This change makes it possible to execute INSERT, UPDATE or DELETE
on tables with row filters. However, given operation is only allowed to
modified rows that are within a row filter.

In case of INSERT when row is matching a row filter then query will
fail.

In case of DELETE and UPDATE rows that are not matching the row filters
are not read at all and so they are not modified or deleted.